Actualizacion de Todo

+++ src/main/java/com/barbersync/api/features/servicio/dto/ServicioResponse.java
@@ -2,13 +2,13 @@ package com.barbersync.api.features.servicio.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
-
 public class ServicioResponse {
-
     private Integer id;
     private String nombre;
     private String descripcion;
-    private double precio;
+    private BigDecimal precio;
     private Integer duracionMinuto;
 }
